requirements.txt
fastapi
uvicorn[standard]
openai
python-dotenv
starlette
# 日志系统
pygelf==0.4.0

# 文档处理依赖
PyPDF2>=3.0.0
pdfplumber>=0.9.0
python-docx>=0.8.11
PyMuPDF>=1.23.0
Pillow>=9.0.0
requests>=2.28.0
beautifulsoup4>=4.11.0
lxml>=4.9.0
markdown>=3.4.0
python-multipart
aiofiles>=23.0.0
zipfile38>=0.0.3
# OCR支持
pytesseract>=0.3.10
opencv-python>=4.8.0 